Change around the Python licenses.
The "canonical" Python license actually has multiple licenses inside of it.
Because of how licenseclassifier attempts to find matching text, this appears
to cause low confidence matches with anything other than an *exact* match.

The first license within the Python license is the Python Foundation License,
which we'll treat as the "Python-2.0.txt" file; and include the
"full" Python license as Python-2.0-complete for exact matching.

In four recent cases this has come up, this now correctly identifies that the
license text contains the Python-2.0 license.

PiperOrigin-RevId: 216559486