use Caller() to find the licenses.db file instead of GOPATH
Instead of looking for the files relative to GOPATH, instead we look for
files relative to the directory file_system_resources.go is in, unless an
absolute path is specified.

This will work under traditional GOPATH and `go mod` scenarios.  It doesn't
help with binaries that are built and the source code moved or removed.