Add new testing scenarios and testing functionality.
A collection of scenarios for existing bugs that were already fixed by the
classifier, plus fixes for two issues that had not been resolved.

SPDX suggests that http and https be treated identically. This change is mapped
into the tokenizer now.

For license headers, users often search-and-replace text for one license and
insert a different license's name. In situations like this, there is a high
percentage match to the original license header. Similar to not changing LGPL
into GPL, we should not transform an Apache header into BSD, or vice versa, but
that was detected in some cases. This fixes that scenario.

Added sorting of license names before verifying against the expected
results to allow stability in testing.