Fix an error in the classifier where license text following END OF TERMS AND CONDITIONS wasn't being removed.
I had been going back and forth between removing this text in the templates or
trimming it programmatically. After realizing we intend to store pristine
license templates for other reasons, pruning the reference content is no longer
a good idea.

This was the root cause of the misfirings of "inconsistency not previously
detected" warnings by the new classifier, since the (usually Apache) license
was not being detected in the license file. Covered by a unit test and a repro
scenario from debugging.