Fix a scoring bug detected in bulk parallel testing.
A variant of the LGPL text bypassed the version change detector because the
phrasing of the version clause is in the wrong location. The layout of the
license happens to cause LGPL-3.0 and LGPL-2.1 to detect at exactly the same
score. Due to the sort and filtering used, the output of the last step of the
classifier was non-deterministic and caused the regression test to flakily
fail.

This fixes the flakiness by causing both licenses to be reported. This
does NOT fix the root cause, since one of the detections needs to be
eliminated. Introducing a template to handle this case doesn't work,
because the version ambiguity issue would spill over into well-formed
files because they would have the "out of order" phrase issue relative
to the "odd" template.