identity_license: prevent send on closed channel
The analyze function calls Done on the waitgroup and then opens up
a new slot by writing into the `task` channel. If this is the last
goroutine running, this can cause an issue: there is a goroutine
running that is waiting for the waitgroup to yield. After this it
closes the `task` channel. If it gets closed before the
`analyze` goroutine manages to send true into the channel, a panic
can occur.

Ensuring `wg.Done()` is called only after `true` is sent into the
channel suitably resolves this.