Change the public API to use []byte instead of string.
The original thinking was that string was a better choice because it clearly
indicated that the classifier wouldn't modify the input data, but it comes with
some runtime cost in that if the caller has a byte slice and just wanted to
scan part of it, they'd need to make a copy to get a string.

Having the API be []byte and documenting that it doesn't modify the data is the
better choice. The currenty implementation doesn't currently realize this
efficiency because it immediately makes a string for tokenization purposes,
but this can be fixed in the future. Changing the API signature is not so
easily accomplished.