Added 3 headers from https://g3doc.corp.google.com/company/teams/opensource/releasing/licenses.md.
I've added the one labeled as BSD-style header in that doc as BSD-3-Clause.header.txt, because the text links to https://developers.google.com/open-source/licenses/bsd, which matches BSD-3-Clause.

I've added the one labeled as MIT header in that doc as MIT.header.txt.

I've added the one labeled GNU GPL header in that doc as GPL-2.0.header_f.txt, since it is the same as the first two paragraphs in GPL-2.0.header_e.txt.

PiperOrigin-RevId: 351922601