Fixes context handling to handle completion better
Changed the usage of contexts so that when the context is done, either
by cancelation or timeout, the classifier finishes the files underway
and abandons any remaining work.

Tested this using a large file and a short context deadline, verifying
that the initial set of files get handled, and subsequent files are
properly abandoned since the context is completed.