Improve matching against licenses that have exception text attached to them.
Remove the trimming of extraneous text feature. It blurs the relationship
between what the input license files look like vs what the classifier actually
uses and doesn't provide value. Fixing this will make it easier to introduce
templates that can match "GPL with exception" licenses since optional text like
the GPL examples won't confound the classifier.

With this change, we need to introduce templates that include the text that was
previously removed, and in the Apache case, create a template of the text
resulting from the trimming that is no longer done.

With these changes, the recognition of exception-based licenses in the corpus
is greatly improved. This also fixes tplicense_test.go back to the way it was
prior to the work on the v2 classifier. The Apache file is recognized as the
pristine Apache file now, rather than being a concatenation of the core Apache
text plus the detection of the header that was in the trimmed text.

PiperOrigin-RevId: 366844535