use a map instead of a slice for the fuseRange filter. It looks like it wants to do some form of random access based on SrcStart / SrcEnd, so this might be a better fit.
<<<
b/196234339 was the problem

There's an off-by one that I don't entirely understand so I'm trying to split the difference here.

The filter variable is currently a slice and we're seeing an off-by-one in production with the amd_vulkan/LICENSE:

```
panic: runtime error: index out of range [3] with length 3 goroutine 1646 [running]:
```

Given that SrcStart / SrcEnd appear to be positions in the text file and the `i` variable seems to move between that range, it seemed natural to replace filter with a map of indexes instead of a slice...this way we can preserve the somewhat random-access pattern that appears to be happening but avoid any range errors.

>>>

PiperOrigin-RevId: 390415408