Fix the induced phrases search to not trigger on modified URLs.
URLs are stored as a single token by the classifier, meaning they can only
introduce 1 error in exact matching, which is good for approximate matching,
since a lengthy URL that is changed would otherwise introduce additional
errors.

However, this doesn't work well with the induced phrases check since a modified
URL would get stored as a insert/delete pair that might contain the triggering
word (e.g. "apache"). This meant that a license that had a different Apache URL
from our pristine copies would get rejected because it "introduced" the word
"apache"

This fixes the logic to not trigger an induced phrase condition if the insert
is paired with a delete that also contains the induced phrase, since this means
it did exist in the document after all. Diagnostic logging for the diffing
phase now includes output to help triage these conditions.

This proved very useful in identifying older Apache licenses that were
incorrectly rejected and sometimes barely matching non-applicable licenses.

PiperOrigin-RevId: 407908906