Fork the licenses content into a new assets directory to more flexibly support content for matching.
Rather than encoding both the type of file and variant name in a single file,
the types are now the top-level directory, the common name of the asset is a
subdirectory, and each file in the directory is a variant that matches the
asset. This makes populating content easier to reason about and allows for easy
extensibility of the type of content that is matched.

The code that uses this new arrangement will be a follow-up CL, since this CL
is purely a branching effort to put content in a new configuration.

PiperOrigin-RevId: 427532245